sdfsdf

+++ app/Http/Controllers/IdeaController.php
@@ -28,18 +28,31 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea) {
 
+        if (auth()->user()->id != $idea->user_id) {
+            abort(404);
+        }
+
         $editing = true;
 
         return view("ideas.show", compact("idea", "editing"));
     }
 
     public function destroy(Idea $idea) {
+
+        if (auth()->user()->id != $idea->user_id) {
+            abort(404);
+        }
+
         $idea->delete();
 
         return redirect()->route("dashboard")->with("success", "Idea Deleted Successfully!");
     }
 
     public function update(Idea $idea) {
+
+        if (auth()->user()->id != $idea->user_id) {
+            abort(404);
+        }
 
         $validated = request()->validate([
             "content" => "required|min:3|max:240"
